Adding svnmerge.py, in it's initial form
This script was inspired by the svnmerge.py tool from orcaware

However, it has several enhancements:
- it works with python3
- it doesn't require the init step the older tool requires
- it uses svn:mergeinfo property right away, so it's compatible
with TortoisSVN, for example
- it uses an interactive menu, inspired from git --interactive
- holds branch merge information into a special property, so one
could invoke this tool with a ? to get a list of known branches.
Otherways any branch could be specified on the command line, and
it'll be added to the known branches list.